Fix version compatibility test on windows
